Changed ints for doubles

+++ src/solver/LinearEquationSolver.java
@@ -1,28 +1,25 @@
 package solver;
 
-
-import exceptions.WrongMatrixSizeException;
-
 public class LinearEquationSolver implements Solver {
 
-    private final int[][] a;
-    private final int[] b;
+    private final double[][] a;
+    private final double[] b;
 
-    public LinearEquationSolver(int[][] a, int[] b) {
+    public LinearEquationSolver(double[][] a, double[] b) {
         this.a = a;
         this.b = b;
     }
 
     @Override
     public double[] solve() {
-        int mainDeterminant = countDeterminant(a);
+        double mainDeterminant = countDeterminant(a);
 
         if (mainDeterminant != 0) {
-            int firstVariableDeterminant = countDeterminant(new int[][]{
+            double firstVariableDeterminant = countDeterminant(new double[][]{
                     {b[0], a[0][1]},
                     {b[1], a[1][1]}
             });
-            int secondVariableDeterminant = countDeterminant(new int[][]{
+            double secondVariableDeterminant = countDeterminant(new double[][]{
                     {a[0][0], b[0]},
                     {a[1][0], b[1]}
             });
@@ -36,7 +33,7 @@ public class LinearEquationSolver implements Solver {
         return null;
     }
 
-    private int countDeterminant(int[][] matrix) {
+    private double countDeterminant(double[][] matrix) {
         return matrix[0][0] * matrix[1][1] - matrix[0][1] * matrix[1][0];
     }
 }
